move step bar to own file
